Moderate API restructure and cleanup - is_move_valid() -> validate_move() and valid_moves() -> valid_ends()
This commit arose out of confusion about where to validate promotion information, which raised the lack of clarity in what a "move" was. As a tuple of (start, end, promotion_data), should Game.valid_moves() cover the range of promotion options where possible? Also should validation of this not belong in Game.is_move_valid() and not in Game.move()?

Accordingly, valid_moves() became valid_ends(), returning a set of end squares that the piece at start could move to (with the side bonus of cleaning up test code), as distinct from valid moves - i.e. no ambiguity over whether or not it should generate promotion options.

Along with this, is_move_valid(), from which a True result was not a sufficient indicator of move validity (it was just a check for end-square validity - it didn't check things like "does this move require promotion data"), has become validate_move(), which should be a NOP if the move is valid, but will raise any of a number of Exceptions in the case of an invalid move. This doesn't particularly feel ideal, but suits the nature of the method (primarily input validation) so I'm going with it.